requirements-reports.txt
requests==2.32.3
python-dotenv==1.0.1
supabase==2.9.0
Pillow==10.4.0
reportlab==4.2.2

